+++ backend/garpix_menu/serializers.py
@@ -1,3 +1,4 @@
+from rest_framework.fields import SerializerMethodField
 from rest_framework.serializers import ModelSerializer
 from .models import MenuItem
 
@@ -7,3 +8,24 @@ class MenuItemSerializer(ModelSerializer):
     class Meta:
         model = MenuItem
         fields = '__all__'
+
+
+class MenuItemWithChildrenSerializer(ModelSerializer):
+    children = SerializerMethodField()
+
+    def get_children(self, instance):
+        items = MenuItem.objects.filter(parent=instance)
+        return MenuItemWithChildrenSerializer(items, many=True).data
+
+    def get_field_names(self, declared_fields, info):
+        expanded_fields = super().get_field_names(declared_fields, info)
+
+        if getattr(self.Meta, 'extra_fields', None):
+            return expanded_fields + self.Meta.extra_fields
+        else:
+            return expanded_fields
+
+    class Meta:
+        model = MenuItem
+        exclude = ('title_for_admin',)
+        extra_fields = ['children']